wip: some more repackaging
